requirements.txt
tensorflow_hub
tensorflow_datasets
pyyaml
torch
datascope==0.0.3
torch
pyarrow
tqdm
pandas
xgboost
seaborn